chore: refactor snapshots, add check for hcp_fs_edits
