+++ python/qx_utilities/general/process.py
@@ -31,6 +31,7 @@ import general.commands_support as gcs
 from processing import fs, simple, workflow, dwi, fsl, rapidtide
 from general.bids import map_nii2bids
 from general import extensions
+from general.parsing import true_or_false as torf, flag, is_none
 
 # pipelines imports
 from hcp import process_hcp
@@ -94,49 +95,6 @@ def procResponse(r):
         return (r, ("Unknown", "Unknown", None))
 
 
-def torf(s):
-    """
-    ``torf(s)``
-
-    First checks if string is "None", 'none', or "NONE" and returns
-    None, then Checks if s is any of the possible true strings: "True", "true",
-    or "TRUE" and returns a boolean result of the check.
-    """
-    if s in ["None", "none", "NONE"]:
-        return None
-    else:
-        return s in ["True", "true", "TRUE", "yes", "Yes", "YES", True]
-
-
-def flag(f):
-    """
-    ``flag(f)``
-
-    Converts a flag (f) passed as a string to a boolean.
-    """
-
-    if type(f) == bool:
-        return f
-    elif f in ["True", "true", "TRUE", "yes", "Yes", "YES"]:
-        return True
-    else:
-        return False
-
-
-def isNone(s):
-    """
-    ``isNone(s)``
-
-    Check if the string is "" and returns None, otherwise
-    returns the passed string.
-    """
-
-    if s in [""]:
-        return None
-    else:
-        return s
-
-
 def update_options(session, options):
     """
     ``update_options(session, options)``
@@ -187,7 +145,7 @@ arglist = [
     ["batchfile", "", str],
     ["sessions", "", str],
     ["sessionsfolder", "", os.path.abspath],
-    ["logfolder", "", isNone],
+    ["logfolder", "", is_none],
     ["logtag", "", str,],
     ["overwrite", "no", torf,],
     ["parsubjects", "1", int,],
@@ -197,7 +155,7 @@ arglist = [
     ["datainfo", "False", torf,],
     ["printoptions", "False", torf,],
     ["filter", "", str,],
-    ["script", "", isNone,],
+    ["script", "", is_none,],
     ["sessionid", "", str,],
     ["sessionids", "", str,],
     ["# ---- Preprocessing options"],
@@ -210,10 +168,10 @@ arglist = [
     ["bold_nuisance", "m,V,WM,WB,1d", str,],
     ["bolds", "all", str,],
     ["boldname", "bold", str,],
-    ["qx_nifti_tail", "", isNone,],
-    ["qx_cifti_tail", "", isNone,],
-    ["nifti_tail", "", isNone,],
-    ["cifti_tail", "", isNone,],
+    ["qx_nifti_tail", "", is_none,],
+    ["qx_cifti_tail", "", is_none,],
+    ["nifti_tail", "", is_none,],
+    ["cifti_tail", "", is_none,],
     ["bold_prefix", "", str,],
     ["bold_variant", "", str,],
     ["img_suffix", "", str,],
@@ -265,14 +223,14 @@ arglist = [
     ["lopass_filter", "0.09", float,],
     ["hipass_do", "nuisance", str,],
     ["lopass_do", "nuisance,movement,events,task", str,],
-    ["omp_threads", "", isNone,],
+    ["omp_threads", "", is_none,],
     ["framework_path", "", str,],
     ["wb_command_path", "", str,],
     ["print_command", "no", str,],
     ["# ---- scheduler options"],
     ["scheduler", "local", str,],
-    ["scheduler_environment", "", isNone,],
-    ["scheduler_workdir", "", isNone,],
+    ["scheduler_environment", "", is_none,],
+    ["scheduler_workdir", "", is_none,],
     ["scheduler_sleep", "1", float,],
     ["# --- general HCP options"],
     ["hcp_processing_mode", "HCPStyleData", str,],
@@ -289,42 +247,42 @@ arglist = [
     ["hcp_hiresmesh", "164", int,],
     ["hcp_bold_res", "2", str,],
     ["hcp_grayordinatesres", "2", int,],
-    ["hcp_surfatlasdir", "", isNone,],
-    ["hcp_grayordinatesdir", "", isNone,],
-    ["hcp_subcortgraylabels", "", isNone,],
-    ["hcp_refmyelinmaps", "", isNone,],
+    ["hcp_surfatlasdir", "", is_none,],
+    ["hcp_grayordinatesdir", "", is_none,],
+    ["hcp_subcortgraylabels", "", is_none,],
+    ["hcp_refmyelinmaps", "", is_none,],
     ["hcp_regname", "MSMSulc", str,],
     ["hcp_cifti_tail", "_Atlas", str,],
     ["hcp_bold_variant", "", str,],
-    ["additional_bolds", "", isNone,],
+    ["additional_bolds", "", is_none,],
     ["hcp_nifti_tail", "", str,],
-    ["hcp_config", "", isNone,],
+    ["hcp_config", "", is_none,],
     ["# --- hcp_pre_freesurfer options"],
     ["hcp_brainsize", "150", int],
     ["hcp_t1samplespacing", "NONE", str],
     ["hcp_t2samplespacing", "NONE", str],
     ["hcp_gdcoeffs", "", str],
     ["hcp_bfsigma", "", str],
-    ["hcp_avgrdcmethod", "", isNone],
+    ["hcp_avgrdcmethod", "", is_none],
     ["hcp_unwarpdir", "z", str],
-    ["hcp_echodiff", "", isNone],
+    ["hcp_echodiff", "", is_none],
     ["hcp_topupconfig", "NONE", str],
     ["hcp_prefs_custombrain", "", str],
-    ["hcp_prefs_template_res", "", isNone],
-    ["hcp_sephaseneg", "", isNone],
-    ["hcp_sephasepos", "", isNone],
-    ["hcp_seechospacing", "", isNone],
-    ["hcp_seunwarpdir", "", isNone],
+    ["hcp_prefs_template_res", "", is_none],
+    ["hcp_sephaseneg", "", is_none],
+    ["hcp_sephasepos", "", is_none],
+    ["hcp_seechospacing", "", is_none],
+    ["hcp_seunwarpdir", "", is_none],
     ["hcp_bold_smoothFWHM", "2", int],
-    ["hcp_prefs_t1template", "", isNone],
-    ["hcp_prefs_t1templatebrain", "", isNone],
-    ["hcp_prefs_t1template2mm", "", isNone],
-    ["hcp_prefs_t2template", "", isNone],
-    ["hcp_prefs_t2templatebrain", "", isNone],
-    ["hcp_prefs_t2template2mm", "", isNone],
-    ["hcp_prefs_templatemask", "", isNone],
-    ["hcp_prefs_template2mmmask", "", isNone],
-    ["hcp_prefs_fnirtconfig", "", isNone],
+    ["hcp_prefs_t1template", "", is_none],
+    ["hcp_prefs_t1templatebrain", "", is_none],
+    ["hcp_prefs_t1template2mm", "", is_none],
+    ["hcp_prefs_t2template", "", is_none],
+    ["hcp_prefs_t2templatebrain", "", is_none],
+    ["hcp_prefs_t2template2mm", "", is_none],
+    ["hcp_prefs_templatemask", "", is_none],
+    ["hcp_prefs_template2mmmask", "", is_none],
+    ["hcp_prefs_fnirtconfig", "", is_none],
     ["# --- hcp_freesurfer options"],
     ["hcp_fs_seed", "", str],
     ["hcp_fs_existing_session", "False", torf],
@@ -340,10 +298,10 @@ arglist = [
     ["# --- hcp_fmri_volume options"],
     ["hcp_bold_biascorrection", "NONE", str],
     ["hcp_bold_usejacobian", "", str],
-    ["hcp_bold_echospacing", "", isNone],
+    ["hcp_bold_echospacing", "", is_none],
     ["hcp_bold_sbref", "NONE", str],
-    ["hcp_bold_dcmethod", "", isNone],
-    ["hcp_bold_echodiff", "", isNone],
+    ["hcp_bold_dcmethod", "", is_none],
+    ["hcp_bold_echodiff", "", is_none],
     ["hcp_bold_unwarpdir", "y", str],
     ["hcp_bold_gdcoeffs", "NONE", str],
     ["hcp_bold_doslicetime", "False", torf],
@@ -354,176 +312,176 @@ arglist = [
     ["hcp_bold_seimg", "independent", str],
     ["hcp_bold_refreg", "", str],
     ["hcp_bold_mask", "", str],
-    ["hcp_bold_sephaseneg", "", isNone],
-    ["hcp_bold_sephasepos", "", isNone],
-    ["hcp_bold_seechospacing", "", isNone],
-    ["hcp_bold_seunwarpdir", "", isNone],
-    ["hcp_bold_topupconfig", "", isNone],
+    ["hcp_bold_sephaseneg", "", is_none],
+    ["hcp_bold_sephasepos", "", is_none],
+    ["hcp_bold_seechospacing", "", is_none],
+    ["hcp_bold_seunwarpdir", "", is_none],
+    ["hcp_bold_topupconfig", "", is_none],
     ["hcp_bold_preregistertool", "", str],
     ["hcp_bold_dof", "", str],
     ["hcp_bold_stcorrdir", "", str],
     ["hcp_bold_stcorrint", "", str],
     ["hcp_wb_resample", "", flag],
-    ["hcp_echo_te", "", isNone],
+    ["hcp_echo_te", "", is_none],
     ["longitudinal", None, flag],
     ["# --- hcp_diffusion options"],
     ["hcp_dwi_echospacing", "", str],
     ["hcp_dwi_phasepos", "PA", str],
     ["hcp_dwi_gdcoeffs", "NONE", str],
-    ["hcp_dwi_dof", "", isNone],
-    ["hcp_dwi_b0maxbval", "", isNone],
-    ["hcp_dwi_combinedata", "", isNone],
-    ["hcp_dwi_extraeddyarg", "", isNone],
-    ["hcp_dwi_name", "", isNone],
+    ["hcp_dwi_dof", "", is_none],
+    ["hcp_dwi_b0maxbval", "", is_none],
+    ["hcp_dwi_combinedata", "", is_none],
+    ["hcp_dwi_extraeddyarg", "", is_none],
+    ["hcp_dwi_name", "", is_none],
     ["hcp_nogpu", None, flag],
-    ["hcp_high_myelin", "", isNone],
+    ["hcp_high_myelin", "", is_none],
     ["hcp_dwi_selectbestb0", None, flag],
     ["hcp_dwi_even_slices", None, flag],
-    ["hcp_dwi_topupconfig", "", isNone],
-    ["hcp_dwi_posdata", "", isNone],
-    ["hcp_dwi_negdata", "", isNone],
+    ["hcp_dwi_topupconfig", "", is_none],
+    ["hcp_dwi_posdata", "", is_none],
+    ["hcp_dwi_negdata", "", is_none],
     ["hcp_dwi_dummy_bval_bvec", None, flag],
     ["# --- general hcp_icafix, hcp_post_fix, hcp_reapply_fix, hcp_msmall, hcp_dedrift_and_resample options"],
-    ["hcp_icafix_bolds", "", isNone],
-    ["hcp_icafix_highpass", "", isNone],
-    ["hcp_matlab_mode", "", isNone],
-    ["hcp_icafix_domotionreg", "", isNone],
-    ["hcp_icafix_deleteintermediates", "", isNone],
-    ["hcp_icafix_fallbackthreshold", "", isNone],
-    ["hcp_icafix_parallel_limit", "", isNone],
-    ["hcp_clean_substring", "", isNone],
+    ["hcp_icafix_bolds", "", is_none],
+    ["hcp_icafix_highpass", "", is_none],
+    ["hcp_matlab_mode", "", is_none],
+    ["hcp_icafix_domotionreg", "", is_none],
+    ["hcp_icafix_deleteintermediates", "", is_none],
+    ["hcp_icafix_fallbackthreshold", "", is_none],
+    ["hcp_icafix_parallel_limit", "", is_none],
+    ["hcp_clean_substring", "", is_none],
     ["# --- hcp_icafix options"],
-    ["hcp_icafix_model", "", isNone],
-    ["hcp_icafix_threshold", "", isNone],
+    ["hcp_icafix_model", "", is_none],
+    ["hcp_icafix_threshold", "", is_none],
     ["hcp_icafix_postfix", "True", torf],
-    ["hcp_icafix_processingmode", "", isNone],
-    ["hcp_icafix_icadim_mode", "", isNone],
-    ["hcp_reuse_existing_ica", "", isNone],
-    ["hcp_fix_backup", "", isNone],
-    ["hcp_t1wtemplatebrain", "", isNone],
-    ["hcp_ica_method", "", isNone],
-    ["hcp_vol_wisharts", "", isNone],
-    ["hcp_cifti_wisharts", "", isNone],
-    ["hcp_icadim_mode", "", isNone],
+    ["hcp_icafix_processingmode", "", is_none],
+    ["hcp_icafix_icadim_mode", "", is_none],
+    ["hcp_reuse_existing_ica", "", is_none],
+    ["hcp_fix_backup", "", is_none],
+    ["hcp_t1wtemplatebrain", "", is_none],
+    ["hcp_ica_method", "", is_none],
+    ["hcp_vol_wisharts", "", is_none],
+    ["hcp_cifti_wisharts", "", is_none],
+    ["hcp_icadim_mode", "", is_none],
     ["hcp_legacy_fix", "", flag],
     ["hcp_icafix_concatenate_only", "", flag],
     ["# --- hcp_post_fix options"],
-    ["hcp_postfix_dualscene", "", isNone],
-    ["hcp_postfix_singlescene", "", isNone],
+    ["hcp_postfix_dualscene", "", is_none],
+    ["hcp_postfix_singlescene", "", is_none],
     ["hcp_postfix_reusehighpass", "True", torf],
     ["# --- hcp_reapply_fix options"],
     ["hcp_icafix_regname", "NONE", str],
     ["# --- hcp_msmall options options"],
-    ["hcp_msmall_bolds", "", isNone],
+    ["hcp_msmall_bolds", "", is_none],
     ["hcp_msmall_outfmriname", "rfMRI_REST", str],
-    ["hcp_msmall_templates", "", isNone],
+    ["hcp_msmall_templates", "", is_none],
     ["hcp_msmall_outregname", "MSMAll_InitialReg", str],
-    ["hcp_msmall_procstring", "", isNone],
+    ["hcp_msmall_procstring", "", is_none],
     ["hcp_msmall_resample", "True", torf],
-    ["hcp_msmall_myelin_target", "", isNone],
-    ["hcp_msmall_module_name", "", isNone],
-    ["hcp_msmall_iteration_modes", "", isNone],
-    ["hcp_msmall_method", "", isNone],
+    ["hcp_msmall_myelin_target", "", is_none],
+    ["hcp_msmall_module_name", "", is_none],
+    ["hcp_msmall_iteration_modes", "", is_none],
+    ["hcp_msmall_method", "", is_none],
     ["hcp_msmall_use_migp", "", flag],
-    ["hcp_msmall_ica_dim", "", isNone],
-    ["hcp_msmall_low_sica_dims", "", isNone],
+    ["hcp_msmall_ica_dim", "", is_none],
+    ["hcp_msmall_low_sica_dims", "", is_none],
     ["hcp_msmall_vn", "", flag],
-    ["hcp_msmall_reg_conf_path", "", isNone],
-    ["hcp_msmall_reg_vars", "", isNone],
-    ["hcp_msmall_rsn_template", "", isNone],
-    ["hcp_msmall_rsn_weights", "", isNone],
-    ["hcp_msmall_topography_roi", "", isNone],
-    ["hcp_msmall_topography_target", "", isNone],
+    ["hcp_msmall_reg_conf_path", "", is_none],
+    ["hcp_msmall_reg_vars", "", is_none],
+    ["hcp_msmall_rsn_template", "", is_none],
+    ["hcp_msmall_rsn_weights", "", is_none],
+    ["hcp_msmall_topography_roi", "", is_none],
+    ["hcp_msmall_topography_target", "", is_none],
     ["hcp_msmall_no_ind_mean", "", flag],
-    ["hcp_msmall_start_frame", "", isNone],
-    ["hcp_msmall_end_frame", "", isNone],
+    ["hcp_msmall_start_frame", "", is_none],
+    ["hcp_msmall_end_frame", "", is_none],
     ["# --- hcp_dedrift_and_resample options"],
     ["hcp_resample_concatregname", "MSMAll", str],
-    ["hcp_resample_regname", "", isNone],
-    ["hcp_resample_reg_files", "", isNone],
+    ["hcp_resample_regname", "", is_none],
+    ["hcp_resample_reg_files", "", is_none],
     ["hcp_resample_maps", "sulc,curvature,corrThickness,thickness", str],
     ["hcp_resample_myelinmaps", "MyelinMap,SmoothedMyelinMap", str],
-    ["hcp_resample_dontfixnames", "", isNone],
-    ["hcp_resample_inregname", "", isNone],
-    ["hcp_resample_use_ind_mean", "", isNone],
-    ["hcp_resample_extractnames", "", isNone],
-    ["hcp_resample_extractextraregnames", "", isNone],
-    ["hcp_resample_extractvolume", "", isNone],
+    ["hcp_resample_dontfixnames", "", is_none],
+    ["hcp_resample_inregname", "", is_none],
+    ["hcp_resample_use_ind_mean", "", is_none],
+    ["hcp_resample_extractnames", "", is_none],
+    ["hcp_resample_extractextraregnames", "", is_none],
+    ["hcp_resample_extractvolume", "", is_none],
     ["# --- hcp_task_fmri_analysis options"],
-    ["hcp_task_lvl1tasks", "", isNone],
-    ["hcp_task_lvl1fsfs", "", isNone],
-    ["hcp_task_lvl2task", "", isNone],
-    ["hcp_task_lvl2fsf", "", isNone],
-    ["hcp_task_summaryname", "", isNone],
-    ["hcp_task_confound", "", isNone],
-    ["hcp_bold_final_smoothFWHM", "", isNone],
-    ["hcp_task_highpass", "", isNone],
-    ["hcp_task_lowpass", "", isNone],
-    ["hcp_task_procstring", "", isNone],
-    ["hcp_task_parcellation", "", isNone],
-    ["hcp_task_parcellation_file", "", isNone],
+    ["hcp_task_lvl1tasks", "", is_none],
+    ["hcp_task_lvl1fsfs", "", is_none],
+    ["hcp_task_lvl2task", "", is_none],
+    ["hcp_task_lvl2fsf", "", is_none],
+    ["hcp_task_summaryname", "", is_none],
+    ["hcp_task_confound", "", is_none],
+    ["hcp_bold_final_smoothFWHM", "", is_none],
+    ["hcp_task_highpass", "", is_none],
+    ["hcp_task_lowpass", "", is_none],
+    ["hcp_task_procstring", "", is_none],
+    ["hcp_task_parcellation", "", is_none],
+    ["hcp_task_parcellation_file", "", is_none],
     ["hcp_task_vba", None, flag],
     ["# --- hcp_asl options"],
-    ["hcp_asl_mtname", "", isNone],
-    ["hcp_asl_territories_atlas", "", isNone],
-    ["hcp_asl_territories_labels", "", isNone],
-    ["hcp_asl_cores", "", isNone],
-    ["hcp_asl_interpolation", "", isNone],
+    ["hcp_asl_mtname", "", is_none],
+    ["hcp_asl_territories_atlas", "", is_none],
+    ["hcp_asl_territories_labels", "", is_none],
+    ["hcp_asl_cores", "", is_none],
+    ["hcp_asl_interpolation", "", is_none],
     ["hcp_asl_use_t1", None, flag],
     ["hcp_asl_nobandingcorr", None, flag],
-    ["hcp_asl_stages", None, isNone],
-    ["hcp_asl_ntis", None, isNone],
-    ["hcp_asl_tis", None, isNone],
-    ["hcp_asl_rpts", None, isNone],
-    ["hcp_asl_bolus", None, isNone],
-    ["hcp_asl_slicedt", None, isNone],
-    ["hcp_asl_sliceband", None, isNone],
-    ["hcp_asl_te", None, isNone],
-    ["hcp_asl_tail_discard_vols", None, isNone],
-    ["hcp_asl_ibf", None, isNone],
+    ["hcp_asl_stages", None, is_none],
+    ["hcp_asl_ntis", None, is_none],
+    ["hcp_asl_tis", None, is_none],
+    ["hcp_asl_rpts", None, is_none],
+    ["hcp_asl_bolus", None, is_none],
+    ["hcp_asl_slicedt", None, is_none],
+    ["hcp_asl_sliceband", None, is_none],
+    ["hcp_asl_te", None, is_none],
+    ["hcp_asl_tail_discard_vols", None, is_none],
+    ["hcp_asl_ibf", None, is_none],
     ["# --- hcp_temporal_ica options"],
-    ["hcp_tica_studyfolder", "", isNone],
-    ["hcp_tica_bolds", "", isNone],
+    ["hcp_tica_studyfolder", "", is_none],
+    ["hcp_tica_bolds", "", is_none],
     ["hcp_tica_outfmriname", "rfMRI_REST", str],
-    ["hcp_tica_surfregname", "", isNone],
-    ["hcp_tica_procstring", "", isNone],
-    ["hcp_outgroupname", "", isNone],
-    ["hcp_tica_timepoints", "", isNone],
-    ["hcp_tica_num_wishart", "", isNone],
-    ["hcp_tica_mrfix_concat_name", "", isNone],
-    ["hcp_tica_icamode", "", isNone],
-    ["hcp_tica_precomputed_clean_folder", "", isNone],
-    ["hcp_tica_precomputed_fmri_name", "", isNone],
-    ["hcp_tica_precomputed_group_name", "", isNone],
-    ["hcp_tica_extra_output_suffix", "", isNone],
-    ["hcp_tica_pca_out_dim", "", isNone],
-    ["hcp_tica_pca_internal_dim", "", isNone],
-    ["hcp_tica_migp_resume", "", isNone],
-    ["hcp_tica_sicadim_iters", "", isNone],
-    ["hcp_tica_sicadim_override", "", isNone],
-    ["hcp_low_sica_dims", "", isNone],
-    ["hcp_tica_reclean_mode", "", isNone],
-    ["hcp_tica_starting_step", "", isNone],
+    ["hcp_tica_surfregname", "", is_none],
+    ["hcp_tica_procstring", "", is_none],
+    ["hcp_outgroupname", "", is_none],
+    ["hcp_tica_timepoints", "", is_none],
+    ["hcp_tica_num_wishart", "", is_none],
+    ["hcp_tica_mrfix_concat_name", "", is_none],
+    ["hcp_tica_icamode", "", is_none],
+    ["hcp_tica_precomputed_clean_folder", "", is_none],
+    ["hcp_tica_precomputed_fmri_name", "", is_none],
+    ["hcp_tica_precomputed_group_name", "", is_none],
+    ["hcp_tica_extra_output_suffix", "", is_none],
+    ["hcp_tica_pca_out_dim", "", is_none],
+    ["hcp_tica_pca_internal_dim", "", is_none],
+    ["hcp_tica_migp_resume", "", is_none],
+    ["hcp_tica_sicadim_iters", "", is_none],
+    ["hcp_tica_sicadim_override", "", is_none],
+    ["hcp_low_sica_dims", "", is_none],
+    ["hcp_tica_reclean_mode", "", is_none],
+    ["hcp_tica_starting_step", "", is_none],
     ["hcp_tica_stop_after_step", "ComputeTICAFeatures", str],
-    ["hcp_tica_remove_manual_components", "", isNone],
-    ["hcp_tica_fix_legacy_bias", "", isNone],
-    ["hcp_parallel_limit", "", isNone],
-    ["hcp_tica_average_dataset", "", isNone],
-    ["hcp_tica_extract_fmri_name_list", "", isNone],
-    ["hcp_tica_extract_fmri_out", "", isNone],
+    ["hcp_tica_remove_manual_components", "", is_none],
+    ["hcp_tica_fix_legacy_bias", "", is_none],
+    ["hcp_parallel_limit", "", is_none],
+    ["hcp_tica_average_dataset", "", is_none],
+    ["hcp_tica_extract_fmri_name_list", "", is_none],
+    ["hcp_tica_extract_fmri_out", "", is_none],
     ["hcp_tica_config_out", None, flag],
     ["hcp_tica_longitudinal_extract_all", None, flag],
-    ["hcp_longitudinal_subject", None, isNone],
-    ["hcp_longitudinal_sessions", None, isNone],
+    ["hcp_longitudinal_subject", None, is_none],
+    ["hcp_longitudinal_sessions", None, is_none],
     ["# --- hcp_apply_auto_reclean options"],
-    ["hcp_autoreclean_model_folder", "", isNone],
-    ["hcp_autoreclean_model_to_use", "", isNone],
-    ["hcp_autoreclean_vote_threshold", "", isNone],
+    ["hcp_autoreclean_model_folder", "", is_none],
+    ["hcp_autoreclean_model_to_use", "", is_none],
+    ["hcp_autoreclean_vote_threshold", "", is_none],
     ["# --- hcp_make_average_dataset options"],
-    ["hcp_surface_atlas_dir", "", isNone],
-    ["hcp_grayordinates_dir", "", isNone],
-    ["hcp_freesurfer_labels", "", isNone],
-    ["hcp_thickness_regression", "", isNone],
+    ["hcp_surface_atlas_dir", "", is_none],
+    ["hcp_grayordinates_dir", "", is_none],
+    ["hcp_freesurfer_labels", "", is_none],
+    ["hcp_thickness_regression", "", is_none],
     ["hcp_pregradient_smoothing", "1", int],
     ["hcp_mad_regname", "MSMAll", str],
     ["hcp_mad_videen_maps", "corrThickness,thickness,MyelinMap_BC,SmoothedMyelinMap_BC", str],
@@ -544,76 +502,76 @@ arglist = [
     ["run", "run", str],
     ["log", "keep", str],
     ["# --- mice pipelines"],
-    ["voxel_increase", "", isNone],
+    ["voxel_increase", "", is_none],
     ["orientation", "x -y z", str],
     ["no_despike", "", flag],
     ["bias_field_correction", "yes", str],
-    ["melodic_anatfile", "", isNone],
-    ["fix_rdata", "", isNone],
+    ["melodic_anatfile", "", is_none],
+    ["fix_rdata", "", is_none],
     ["fix_threshold", "20", int],
     ["fix_no_motion_cleanup", "", flag],
     ["fix_aggressive_cleanup", "", flag],
     ["mice_highpass", "0.01", str],
     ["mice_lowpass", "0.25", str],
     ["mice_volumes", "900", int],
-    ["flirt_ref", "", isNone],
+    ["flirt_ref", "", is_none],
     ["# --- hcp_long_freesurfer options"],
     ["hcp_longitudinal_template", "base", str],
     ["hcp_no_t2w", "", flag],
-    ["hcp_seed", "", isNone],
+    ["hcp_seed", "", is_none],
     ["hcp_parallel_mode", "BUILTIN", str],
-    ["hcp_fslsub_queue", "", isNone],
-    ["hcp_max_jobs", "", isNone],
-    ["hcp_start_stage", "", isNone],
-    ["hcp_end_stage", "", isNone],
+    ["hcp_fslsub_queue", "", is_none],
+    ["hcp_max_jobs", "", is_none],
+    ["hcp_start_stage", "", is_none],
+    ["hcp_end_stage", "", is_none],
     ["# --- hcp_transmit_bias_individual options"],
-    ["hcp_transmit_mode", "", isNone],
-    ["hcp_gmwm_template", "", isNone],
-    ["hcp_group_corrected_myelin", "", isNone],
-    ["hcp_afi_image", "", isNone],
-    ["hcp_afi_tr_one", "", isNone],
-    ["hcp_afi_tr_two", "", isNone],
-    ["hcp_afi_angle", "", isNone],
-    ["hcp_b1tx_magnitude", "", isNone],
-    ["hcp_b1tx_phase", "", isNone],
-    ["hcp_b1tx_phase_divisor", "", isNone],
-    ["hcp_pt_fmri_names", "", isNone],
-    ["hcp_pt_bbr_threshold", "", isNone],
-    ["hcp_myelin_template", "", isNone],
-    ["hcp_group_uncorrected_myelin", "", isNone],
-    ["hcp_pt_reference_value_file", "", isNone],
-    ["hcp_unproc_t1w_list", "", isNone],
-    ["hcp_unproc_t2w_list", "", isNone],
-    ["hcp_receive_bias_body_coil", "", isNone],
-    ["hcp_receive_bias_head_coil", "", isNone],
-    ["hcp_raw_psn_t1w", "", isNone],
-    ["hcp_raw_nopsn_t1w", "", isNone],
-    ["hcp_transmit_res", "", isNone],
-    ["hcp_myelin_mapping_fwhm", "", isNone],
+    ["hcp_transmit_mode", "", is_none],
+    ["hcp_gmwm_template", "", is_none],
+    ["hcp_group_corrected_myelin", "", is_none],
+    ["hcp_afi_image", "", is_none],
+    ["hcp_afi_tr_one", "", is_none],
+    ["hcp_afi_tr_two", "", is_none],
+    ["hcp_afi_angle", "", is_none],
+    ["hcp_b1tx_magnitude", "", is_none],
+    ["hcp_b1tx_phase", "", is_none],
+    ["hcp_b1tx_phase_divisor", "", is_none],
+    ["hcp_pt_fmri_names", "", is_none],
+    ["hcp_pt_bbr_threshold", "", is_none],
+    ["hcp_myelin_template", "", is_none],
+    ["hcp_group_uncorrected_myelin", "", is_none],
+    ["hcp_pt_reference_value_file", "", is_none],
+    ["hcp_unproc_t1w_list", "", is_none],
+    ["hcp_unproc_t2w_list", "", is_none],
+    ["hcp_receive_bias_body_coil", "", is_none],
+    ["hcp_receive_bias_head_coil", "", is_none],
+    ["hcp_raw_psn_t1w", "", is_none],
+    ["hcp_raw_nopsn_t1w", "", is_none],
+    ["hcp_transmit_res", "", is_none],
+    ["hcp_myelin_mapping_fwhm", "", is_none],
     ["hcp_old_myelin_mapping", "", flag],
     ["# --- fsl_feat options"],
-    ["feat_file", "", isNone],
+    ["feat_file", "", is_none],
     ["# --- fsl_melodic options"],
-    ["input_files", "", isNone],
-    ["melodic_extra_args", "", isNone],
+    ["input_files", "", is_none],
+    ["melodic_extra_args", "", is_none],
     ["# --- rapidtide options"],
-    ["despecklepasses", "", isNone,],
-    ["filterband", "", isNone,],
-    ["searchrange", "", isNone,],
-    ["nprocs", "", isNone,],
+    ["despecklepasses", "", is_none,],
+    ["filterband", "", is_none,],
+    ["searchrange", "", is_none,],
+    ["nprocs", "", is_none,],
     ["nofitfilt", "", flag,],
-    ["similaritymetric", "", isNone,],
-    ["ampthresh", "", isNone,],
-    ["numnull", "", isNone,],
-    ["outputlevel", "", isNone,],
-    ["spatialfilt", "", isNone,],
-    ["simcalcrange", "", isNone,],
-    ["brainmask", "", isNone,],
-    ["graymattermask", "", isNone,],
-    ["whitemattermask", "", isNone,],
-    ["refineexclude", "", isNone,],
+    ["similaritymetric", "", is_none,],
+    ["ampthresh", "", is_none,],
+    ["numnull", "", is_none,],
+    ["outputlevel", "", is_none,],
+    ["spatialfilt", "", is_none,],
+    ["simcalcrange", "", is_none,],
+    ["brainmask", "", is_none,],
+    ["graymattermask", "", is_none,],
+    ["whitemattermask", "", is_none,],
+    ["refineexclude", "", is_none,],
     ["nodenoise", "", flag,],
-    ["rapidtide_extra_args", "", isNone],
+    ["rapidtide_extra_args", "", is_none],
 ]
 
 # Add arguments used in extensions
