+++ python/qx_utilities/hcp/process_hcp.py
@@ -28,7 +28,7 @@ import json
 import general.core as gc
 import processing.core as pc
 import general.img as gi
-import general.utilities as gu
+import general.snapshots as gs
 import general.exceptions as ge
 import nibabel as nib
 import pprint
@@ -1831,7 +1831,46 @@ def hcp_freesurfer(sinfo, options, overwrite=False, thread=0):
             ``hcp_fs_no_conf2hires``     ``no-conf2hires``
             ``hcp_fs_flair``             ``flair``
             ``hcp_nogpu``                ``gpu``
-            ============================ =======================
+            ============================ =======================    
+
+        Running FreeSurfer again after PostFreeSurfer:
+
+            If FreeSurfer has to be run again after PostFreeSurfer, the files 
+            generated during PostFreeSurfer need to be removed to prevent
+            any conflicts. For this reason, when PostFreeSurfer is run, QuNex
+            records the list of files generated during PostFreeSurfer and backs
+            up files known to be modified by PostFreeSurfer. When FreeSurfer
+            is run again, files generated by PostFreeSurfer are removed and the 
+            backed up files are restored.
+
+        Running with FreeSurfer edits:
+
+            Sometimes, after visual inspection of the FreeSurfer outputs, manual
+            edits are needed to correct for misclassified regions. If such edits
+            have been made, they can be applied during the FreeSurfer step in 
+            two ways. The first option is to place the edited files in the 
+            FreeSurfer standard location within the subject's FS folder, set the 
+            ``hcp_fs_existing_session`` parameter to 'TRUE' and specify the 
+            recon-all stages to run using the ``hcp_fs_extra_reconall`` parameter. 
+            This way, the command will run on top of an already existing analysis/
+            subject and will incorporate the edits.
+
+            The second option, if only wm.mgz, brainmask.mgz, and/or aseg.mgz 
+            files have been edited, is to place the edited files in the
+            ``<sessions folder>/inbox/FS_edits`` folder and prepend 
+            ``<session id>_`` to identify them. In this case, the 
+            ``hcp_fs_edits`` parameter should be set to 'TRUE'. QuNex will 
+            detect which edits are present and will set the ``hcp_fs_existing_session``
+            and ``hcp_fs_extra_reconall`` parameters accordingly to incorporate 
+            the edits.
+
+            The third option is to place the edited files in the standard location
+            within the subject's FS folder, and specify which edits to apply 
+            using the ``hcp_fs_edits`` parameter as a comma-separated list of 
+            edits to apply: 'aseg', 'wm', 'brainmask'. In this case, the 
+            ``hcp_fs_existing_session`` parameter will be set to 'TRUE', and
+            ``hcp_fs_extra_reconall`` parameter will be set accordingly to
+            incorporate the specified edits.
 
     Examples:
         Example run from the base study folder with test flag::
@@ -1863,6 +1902,26 @@ def hcp_freesurfer(sinfo, options, overwrite=False, thread=0):
                 --hcp_freesurfer_home=<absolute_path_to_freesurfer_binary> \\
                 --scheduler="SLURM,time=03-24:00:00,cpus-per-task=2,mem-per-cpu=1250,partition=week"
 
+        Run with FreeSurfer edits applied and present in inbox/FS-edits::
+
+            qunex hcp_freesurfer \\
+                --batchfile="processing/batch.txt" \\
+                --sessionsfolder="sessions" \\
+                --sessions="s01,sub-07,sub-12" \\
+                --parsessions="3" \\
+                --overwrite="yes" \\
+                --hcp_fs_edits="TRUE"
+
+        Run with FreeSurfer edits prepared within FS folder and specified explicitly::
+
+            qunex hcp_freesurfer \\
+                --batchfile="processing/batch.txt" \\
+                --sessionsfolder="sessions" \\
+                --sessions="s01,sub-07,sub-12" \\
+                --parsessions="3" \\
+                --overwrite="yes" \\
+                --hcp_fs_edits="wm,brainmask"
+                
         Additional examples::
 
             qunex hcp_freesurfer \\
@@ -1947,45 +2006,52 @@ def hcp_freesurfer(sinfo, options, overwrite=False, thread=0):
 
         # do we have edits specified?
         if options["hcp_fs_edits"] not in ["", "False", "false", "No", "no"]:
-            edited = [e for e in ['aseg', 'wm', 'brainmask'] if e in options["hcp_fs_edits"].lower().split(",")]
-            # -- map files
-            r += "\n---> hcp_fs_edits is set to TRUE, looking for edits files ..."
-            editsfolder = os.path.join(options["sessionsfolder"], "inbox", "FS_edits")
-            editfiles = glob.glob(os.path.join(editsfolder, sinfo["id"] + "_*.mgz"))
-            if editfiles:
-                for efile in editfiles:
-                    fname = os.path.basename(efile).split("_", 1)[1]
-                    edited.append(fname.replace(".mgz",""))
-                    destfile = os.path.join(hcp["FS_folder"], 'mri', fname)
-                    if os.path.exists(destfile):
-                        r += "\n     ... replacing: %s " % (fname)
-                    else:
-                        r += "\n     ... adding: %s " % (fname)
-                    shutil.copy2(efile, destfile)
-            else:
-                r += "\n     ... no edits files found in %s!" % (editsfolder)
 
-            # -- set extra parameters
-            options["hcp_fs_existing_session"] = True
-            if 'wm' in edited:
-                add_extra = ['-autorecon2-wm', '-autorecon3']
-            elif 'aseg' in edited:
-                add_extra = ['-autorecon2-noaseg', '-autorecon3']
-            elif 'brainmask' in edited:
-                add_extra = ['-autorecon-pial', '-autorecon3']
-            else:
-                r += "\n---> ERROR: No edits specified and no edited files found!"
-                r += "\n            If you are processing edits to wm, aseg, or brainmask, please provide the appropriate edit files or list them explicitly in hcp_fs_edits."
-                r += "\n            For other edits, please set hcp_fs_edits to FALSE, and use hcp_fs_existing_session and hcp_fs_extra_reconall parameters."
+            extra = [e.strip() for e in options["hcp_fs_edits"].lower().split(",") if e.strip() not in ['aseg', 'wm', 'brainmask', 'yes', 'true']]
+            if extra:
+                r += "\n---> ERROR: Invalid edits specified in hcp_fs_edits: %s" % ",".join(extra)
                 run = False
-                add_extra = []
 
-            if not hcp['T2w'] in ["", "NONE"]:
-                add_extra.append('-T2pial')
+            else:
+                edited = [e for e in ['aseg', 'wm', 'brainmask'] if e in options["hcp_fs_edits"].lower().split(",")]
+                # -- map files
+                r += "\n---> hcp_fs_edits is set to TRUE, looking for edits files ..."
+                editsfolder = os.path.join(options["sessionsfolder"], "inbox", "FS_edits")
+                editfiles = glob.glob(os.path.join(editsfolder, sinfo["id"] + "_*.mgz"))
+                if editfiles:
+                    for efile in editfiles:
+                        fname = os.path.basename(efile).split("_", 1)[1]
+                        edited.append(fname.replace(".mgz",""))
+                        destfile = os.path.join(hcp["FS_folder"], 'mri', fname)
+                        if os.path.exists(destfile):
+                            r += "\n     ... replacing: %s " % (fname)
+                        else:
+                            r += "\n     ... adding: %s " % (fname)
+                        shutil.copy2(efile, destfile)
+                else:
+                    r += "\n     ... no edits files found in %s!" % (editsfolder)
 
-            add_extra = [p for p in add_extra if p not in options["hcp_fs_extra_reconall"]]
-            if add_extra:
-                options["hcp_fs_extra_reconall"] = "|".join(add_extra + [options["hcp_fs_extra_reconall"]])            
+                # -- set extra parameters
+                options["hcp_fs_existing_session"] = True
+                if 'wm' in edited:
+                    add_extra = ['-autorecon2-wm', '-autorecon3']
+                elif 'aseg' in edited:
+                    add_extra = ['-autorecon2-noaseg', '-autorecon3']
+                elif 'brainmask' in edited:
+                    add_extra = ['-autorecon-pial', '-autorecon3']
+                else:
+                    r += "\n---> ERROR: No edits specified and no edited files found!"
+                    r += "\n            If you are processing edits to wm, aseg, or brainmask, please provide the appropriate edit files or list them explicitly in hcp_fs_edits."
+                    r += "\n            For other edits, please set hcp_fs_edits to FALSE, and use hcp_fs_existing_session and hcp_fs_extra_reconall parameters."
+                    run = False
+                    add_extra = []
+
+                if not hcp['T2w'] in ["", "NONE"]:
+                    add_extra.append('-T2pial')
+
+                add_extra = [p for p in add_extra if p not in options["hcp_fs_extra_reconall"]]
+                if add_extra:
+                    options["hcp_fs_extra_reconall"] = "|".join(add_extra + [options["hcp_fs_extra_reconall"]])            
 
         # -> Key elements
         elements = [
@@ -2112,19 +2178,19 @@ def hcp_freesurfer(sinfo, options, overwrite=False, thread=0):
                     # -> cleanup postfs
                     r += "\n     Found PostFreeSurfer results file: %s" % (post_fs_tfile)
                     r += "\n     Cleaning up PostFreeSurfer results to allow FreeSurfer reprocessing ..."
-                    gu.rollback_snapshot(diff=os.path.join(hcp['snapshots'], "postfreesurfer_diff.txt"), 
+                    gs.rollback_snapshot(diff=os.path.join(hcp['snapshots'], "postfreesurfer_diff.txt"), 
                                         action="delete",
                                         exclude=hcp['snapshots'])
 
                     # -> restore backup
                     r += "\n     Restoring FreeSurfer backup ..."
-                    gu.restore_files(source=os.path.join(hcp['snapshots'], "postfreesurfer_backup"),            
+                    gs.restore_files(source=os.path.join(hcp['snapshots'], "postfreesurfer_backup"),            
                                     target=hcp['base'],
                                     overwrite=True)
                 
                 # --> record freesurfer_start_snapshot
                 r += "\n---> Recording FreeSurfer start snapshot ..."
-                gu.record_snapshot(targetfolder=hcp['base'],
+                gs.record_snapshot(targetfolder=hcp['base'],
                                    outfile=os.path.join(hcp['snapshots'], "freesurfer_start.txt"), 
                                    exclude=hcp['snapshots'])
 
@@ -2186,7 +2252,7 @@ def hcp_freesurfer(sinfo, options, overwrite=False, thread=0):
 
     # -- take freesurfer_end_snapshot
     if options["run"] == "run":
-        gu.compare_snapshots(before=os.path.join(hcp['snapshots'], "freesurfer_start.txt"), 
+        gs.compare_snapshots(before=os.path.join(hcp['snapshots'], "freesurfer_start.txt"), 
                              after=os.path.join(hcp['base']),
                              outfile=os.path.join(hcp['snapshots'], "freesurfer_diff.txt"),
                              exclude=hcp['snapshots'])
@@ -2521,12 +2587,12 @@ def hcp_post_freesurfer(sinfo, options, overwrite=False, thread=0):
                 
                 if not os.path.exists(tfile):
                     # ---> record pre freesurfer snapshot
-                    gu.record_snapshot(targetfolder=hcp['base'], 
+                    gs.record_snapshot(targetfolder=hcp['base'], 
                                        outfile=os.path.join(hcp['snapshots'], "postfreesurfer_start.txt"), 
                                        exclude=hcp['snapshots'])
     
                     # ---> prepare freesurfer backup
-                    gu.backup_files(source=hcp['base'], 
+                    gs.backup_files(source=hcp['base'], 
                                     target=os.path.join(hcp['snapshots'], 'postfreesurfer_backup'),
                                     filelist=['MNINonLinear/T1w.nii.gz',
                                               'MNINonLinear/T1w_restore.nii.gz',
@@ -2604,7 +2670,7 @@ def hcp_post_freesurfer(sinfo, options, overwrite=False, thread=0):
 
     # -- take freesurfer_end_snapshot
     if options["run"] == "run":
-        gu.compare_snapshots(before=os.path.join(hcp['snapshots'], "postfreesurfer_start.txt"), 
+        gs.compare_snapshots(before=os.path.join(hcp['snapshots'], "postfreesurfer_start.txt"), 
                              after=os.path.join(hcp['base']),
                              outfile=os.path.join(hcp['snapshots'], "postfreesurfer_diff.txt"),
                              exclude=hcp['snapshots'])
